refactor: rename tool file to use underscores for consistency
